Move context folder and deactivate the autosave
